walrus: Move API types to separate package
I did this mostly because the linter was complaining (about unused
API types in 'main'), but in the process I realized that splitting
out the API types into an importable package makes sense if, you
know, you want people to actually write clients for your server.

More broadly, this has me wondering if walrus should be moved to its
own repo entirely. walrus-cli is already split out (solely because I
wanted to keep us free of cgo), but it would probably be cleaner to
stick it alongside the walrus server within the same repo.

But if walrus is going to live outside us, then what about user, and
shard, and any other future binaries? Should each binary get its own
repo, leaving us as purely a library? I can see the appeal of that,
especially since Go modules seem to encourage that kind of structure.
There's just one problem: names! It's not clear that, for example,
lukechampine/shard is related to us. But I suppose this problem will
be shared by *anyone* who builds on us, so maybe I should just suck
it up.
